Big rework to enable easy switching of themes.
This update is inspired and based on risotto, a Hugo theme.

Rework on the CSS variables that control the colors of the site.
This allows to quickly switch between palettes (themes)
which share a common structure of 16 colors —8 gradient colors
and 8 accent colors.

Enable a way to pick a random theme (palette) for each page.
The random feature allows the user to get quick visual feedback
of the different palettes accross the site.

Adjust CSS classes to reduce noise and ease the implementation
of multiple palettes.