Rework the library template
Make the template consume a csv file with the details of the books in
the library.
Compute metrics out of that csv and display them in paragraphs.
Construct the html table out of the csv data.

This will reduce by a lot the manual work I need to do to update my
library. Also, showing statistics is very cool.

Also, make the library its own section.