Big rework to implement the new template system introduced in Hugo v0.146.0
See: https://gohugo.io/templates/new-templatesystem-overview

In v0.146.0 Hugo introduced a full re-implementation of the template system.
That change would have made impossible to upgrade Hugo to a higher version
without updating the templates within this theme first.
The Hugo change comes with improvements on how templates work.
The previous lookup order was just too messy and led to weird stuff in the theme.

So, this change addresses that.

* _default folder is removed
* partials and shortcodes folders are renamed to _partials and _shortcodes
* major sections apart from blog are given a proper section/page set of templates
* remove Spanish where possible, and prefer English. I'm still writing in Spanish, of course
* some minor tweaks here and there, since I'm already doing a big change, am I not?