refactor: Big rework to make the blog pick the library ID
This fixes a growing issue that was relying on the _page number_ instead
of the library ID to reference books. For things like cover pictures in
the readings pages, or to cross reference library, book, reading page
and review page the previous logic was using the number assigned to the
book instead of the book id.

That "number assigned" is like the position of the book in a shelf. It's
ok for a while, but then if you move something out of place the whole
system falls apart. I did delete a book from my library database (I
registered it by accident) and the whole thing got broken in my blog.
Fortunately, I catched that one in the development server.

So I proceeded with the fix. Now the layouts look for the ID instead and
while the assigned number is still useful (to get the cross reference
for example) is no longer a breaking change if the book moves from one
shelf to another, because now everything looks at the book ID and not
its position.